EVAKA-3968 Cleaned up attendance child pages
