+++ frontend/packages/employee-frontend/src/components/attendances/AttendanceGroupSelectorPage.tsx
@@ -2,68 +2,20 @@
 //
 // SPDX-License-Identifier: LGPL-2.1-or-later
 
-import React, { Fragment, useEffect, useState } from 'react'
+import React, { Fragment, useContext, useEffect } from 'react'
 import { useParams } from 'react-router-dom'
 import styled from 'styled-components'
 import MetaTags from 'react-meta-tags'
 
-import LocalDate from '@evaka/lib-common/src/local-date'
-import { isFailure, isLoading, isSuccess, Loading, Result } from '~api'
-import { getDaycare, getUnitData, UnitData, UnitResponse } from '~api/unit'
-import Button from '~components/shared/atoms/buttons/Button'
+import { isFailure, isLoading, isSuccess } from '~api'
 import Loader from '~components/shared/atoms/Loader'
 import Title from '~components/shared/atoms/Title'
 import { ContentArea } from '~components/shared/layout/Container'
 import { DefaultMargins } from '~components/shared/layout/white-space'
-import { DaycareGroup } from '~types/unit'
 import { useTranslation } from '~state/i18n'
-
-export const Flex = styled.div`
-  @media screen and (max-width: 1023px) {
-    justify-content: space-between;
-  }
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-`
-
-export const FlexColumn = styled.div`
-  @media screen and (max-width: 1023px) {
-    justify-content: space-between;
-  }
-  display: flex;
-  flex-direction: column;
-  flex-wrap: wrap;
-`
-
-export interface CustomButtonProps {
-  color?: string
-  backgroundColor?: string
-  borderColor?: string
-}
-
-export const CustomButton = styled(Button)<CustomButtonProps>`
-  @media screen and (max-width: 1023px) {
-    margin-bottom: ${DefaultMargins.s};
-    width: calc(50vw - 40px);
-    white-space: normal;
-    height: 64px;
-  }
-
-  @media screen and (min-width: 1024px) {
-    margin-right: ${DefaultMargins.s};
-  }
-  ${(p) => (p.color ? `color: ${p.color};` : '')}
-  ${(p) => (p.backgroundColor ? `background-color: ${p.backgroundColor};` : '')}
-  ${(p) => (p.borderColor ? `border-color: ${p.borderColor};` : '')}
-
-  :hover {
-    ${(p) => (p.color ? `color: ${p.color};` : '')}
-    ${(p) =>
-      p.backgroundColor ? `background-color: ${p.backgroundColor};` : ''}
-  ${(p) => (p.borderColor ? `border-color: ${p.borderColor};` : '')}
-  }
-`
+import { AttendanceUIContext } from '~state/attendance-ui'
+import { getDaycareAttendances, Group } from '~api/attendances'
+import { Flex, CustomButton } from './components'
 
 const AllCapsTitle = styled(Title)`
   text-transform: uppercase;
@@ -78,16 +30,15 @@ const Padding = styled.div`
 export default React.memo(function AttendanceGroupSelectorPage() {
   const { i18n } = useTranslation()
   const { unitId } = useParams<{ unitId: string }>()
-  const [unitData, setUnitData] = useState<Result<UnitData>>(Loading())
-  const [unit, setUnit] = useState<Result<UnitResponse>>(Loading())
 
-  const loading = isLoading(unit) || isLoading(unitData)
+  const { attendanceResponse, setAttendanceResponse } = useContext(
+    AttendanceUIContext
+  )
+
+  const loading = isLoading(attendanceResponse)
 
   useEffect(() => {
-    void getUnitData(unitId, LocalDate.today(), LocalDate.today()).then(
-      setUnitData
-    )
-    void getDaycare(unitId).then(setUnit)
+    void getDaycareAttendances(unitId).then(setAttendanceResponse)
   }, [])
 
   return (
@@ -97,12 +48,11 @@ export default React.memo(function AttendanceGroupSelectorPage() {
       </MetaTags>
 
       {loading && <Loader />}
-      {isFailure(unitData) ||
-        (isFailure(unit) && <div>{i18n.common.loadingFailed}</div>)}
-      {isSuccess(unitData) && isSuccess(unit) && (
+      {isFailure(attendanceResponse) && <div>{i18n.common.loadingFailed}</div>}
+      {isSuccess(attendanceResponse) && (
         <Fragment>
           <Title size={1} centered smaller bold>
-            {unit.data.daycare.name}
+            {attendanceResponse.data.unit.name}
           </Title>
           <Padding>
             <ContentArea paddingHorozontal={'s'} opaque>
@@ -113,9 +63,9 @@ export default React.memo(function AttendanceGroupSelectorPage() {
                 <a href={`attendance/all/coming`}>
                   <CustomButton primary text={i18n.common.all} />
                 </a>
-                {unitData.data.groups.map((elem: DaycareGroup) => (
-                  <a key={elem.id} href={`attendance/${elem.id}/coming`}>
-                    <CustomButton primary key={elem.id} text={elem.name} />
+                {attendanceResponse.data.unit.groups.map((group: Group) => (
+                  <a key={group.id} href={`attendance/${group.id}/coming`}>
+                    <CustomButton primary key={group.id} text={group.name} />
                   </a>
                 ))}
               </Flex>
